requirements.txt
django
djangorestframework
django-cors-headers
requests
mysqlclient
pillow
djangorestframework-simplejwt
djangorestframework-simplejwt[blacklist]
channels
channels-redis
